fix: add slippage to memo for regular and streaming swaps
