+++ src/lib/swapper/swappers/ThorchainSwapper/utils/getQuote/getQuote.ts
@@ -2,7 +2,6 @@ import type { AssetId } from '@shapeshiftoss/caip'
 import { bchAssetId } from '@shapeshiftoss/caip'
 import type { Result } from '@sniptt/monads'
 import { Err, Ok } from '@sniptt/monads'
-import BigNumber from 'bignumber.js'
 import { getConfig } from 'config'
 import qs from 'qs'
 import type { Asset } from 'lib/asset-service'
@@ -18,10 +17,8 @@ import {
   THORCHAIN_AFFILIATE_NAME,
   THORCHAIN_FIXED_PRECISION,
 } from 'lib/swapper/swappers/ThorchainSwapper/utils/constants'
-import { assertIsValidMemo } from 'lib/swapper/swappers/ThorchainSwapper/utils/makeSwapMemo/assertIsValidMemo'
 import { assetIdToPoolAssetId } from 'lib/swapper/swappers/ThorchainSwapper/utils/poolAssetHelpers/poolAssetHelpers'
 import { createTradeAmountTooSmallErr } from 'lib/swapper/utils'
-import { subtractBasisPointAmount } from 'state/slices/tradeQuoteSlice/utils'
 
 import { thorService } from '../thorService'
 
@@ -31,7 +28,6 @@ export const getQuote = async ({
   sellAmountCryptoBaseUnit,
   receiveAddress,
   affiliateBps = '0',
-  slippageBps,
 }: {
   sellAsset: Asset
   buyAssetId: AssetId
@@ -39,7 +35,6 @@ export const getQuote = async ({
   // Receive address is optional for THOR quotes, and will be in case we are getting a quote with a missing manual receive address
   receiveAddress: string | undefined
   affiliateBps: string
-  slippageBps: string
 }): Promise<Result<ThornodeQuoteResponseSuccess, SwapErrorRight>> => {
   const buyPoolId = assetIdToPoolAssetId({ assetId: buyAssetId })
   const sellPoolId = assetIdToPoolAssetId({ assetId: sellAsset.assetId })
@@ -99,27 +94,6 @@ export const getQuote = async ({
       }),
     )
   } else {
-    const memoWithManualSlippage = (() => {
-      const MEMO_PART_DELIMITER = ':'
-      // TODO: Woody you'll need to use expected_amount_out_streaming for streaming swaps
-      const { memo: quotedMemo, expected_amount_out: expectedAmountOut } = data
-      const memoParts = quotedMemo.split(MEMO_PART_DELIMITER)
-
-      const pool = memoParts[1]
-      const address = memoParts[2]
-      const affiliate = memoParts[4]
-      const affiliateBps = memoParts[5]
-
-      const limitWithManualSlippage = subtractBasisPointAmount(
-        expectedAmountOut,
-        slippageBps,
-        BigNumber.ROUND_DOWN,
-      )
-
-      const memo = `s${MEMO_PART_DELIMITER}${pool}${MEMO_PART_DELIMITER}${address}${MEMO_PART_DELIMITER}${limitWithManualSlippage}${MEMO_PART_DELIMITER}${affiliate}${MEMO_PART_DELIMITER}${affiliateBps}`
-      assertIsValidMemo(memo)
-      return memo
-    })()
-    return Ok({ ...data, memo: memoWithManualSlippage })
+    return Ok(data)
   }
 }
